requirements.txt
absl-py==2.0.0
accelerate==0.16.0
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
aiostream==0.5.2
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autocfg==0.0.8
autogluon==0.8.2
autogluon.common==0.8.2
autogluon.core==0.8.2
autogluon.features==0.8.2
autogluon.multimodal==0.8.2
autogluon.tabular==0.8.2
autogluon.timeseries==0.8.2
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bertopic==0.15.0
black==23.9.1
bleach==6.1.0
blessed==1.20.0
blis==0.7.11
boto3==1.28.80
botocore==1.31.80
cachetools==5.3.2
catalogue==2.0.10
catboost==1.2.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cleanlab==2.5.0
click==8.1.7
cloudpickle==2.0.0
cmake==3.27.5
colorama==0.4.6
colorful==0.5.5
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.5
cycler==0.11.0
cymem==2.0.8
Cython==0.29.36
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
easygui==0.98.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
et-xmlfile==1.1.0
evaluate==0.3.0
exceptiongroup==1.1.3
executing==1.2.0
fastai==2.7.13
fastapi==0.104.1
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.18.0
fastprogress==1.0.3
filelock==3.12.4
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gdown==4.7.1
gluoncv==0.10.5.post0
gluonts==0.13.7
google-api-core==2.13.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gpustat==1.1.1
graphviz==0.20.1
grpcio==1.50.0
grpclib==0.4.3
h2==4.1.0
hdbscan==0.8.33
hpack==4.0.0
huggingface-hub==0.17.3
hyperframe==6.0.1
hyperopt==0.2.7
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langcodes==3.3.0
lightgbm==3.3.5
lightning-utilities==0.9.0
lit==17.0.1
llvmlite==0.41.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mlforecast==0.7.3
modAL-python==0.4.2.1
model-index==0.1.11
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nlpaug==1.1.11
nltk==3.8.1
notebook==7.0.5
notebook_shim==0.2.3
nptyping==2.4.1
numba==0.58.0
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.2.3
opencensus==0.11.3
opencensus-context==0.1.3
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.2
openTSNE==1.0.0
openxlab==0.0.28
ordered-set==4.1.0
oss2==2.17.0
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pathy==0.10.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.17.0
plotly-express==0.4.1
portalocker==2.8.2
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
pydantic_core==2.6.3
Pygments==2.16.1
pynndescent==0.5.10
pyparsing==3.1.1
pyrsistent==0.20.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-json-logger==2.0.7
python-utils==3.8.1
pytorch-lightning==1.9.5
pytorch-metric-learning==1.7.3
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
ray==2.3.1
referencing==0.30.2
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.10.3
rsa==4.9
s3transfer==0.7.0
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.11.2
seaborn==0.13.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
seqeval==1.2.2
sigtools==4.0.1
six==1.16.0
skorch==0.9.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
statsforecast==1.4.0
statsmodels==0.14.0
sympy==1.12
synchronicity==0.5.3
tabulate==0.9.0
tbb==2021.10.0
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.10
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.3.3
tqdm==4.65.2
traitlets==5.10.0
transformers==4.26.1
triton==2.0.0
typer==0.9.0
types-certifi==2021.10.8.3
types-python-dateutil==2.8.19.14
types-toml==0.10.8.7
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
umap-learn==0.5.4
uri-template==1.3.0
urllib3==1.26.18
utils==1.0.1
virtualenv==20.24.6
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
window-ops==0.0.14
xgboost==1.7.6
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0
